AMP-2927
git-svn-id: file:///svn/trunk/amp@4721 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ WEB-INF/src/org/digijava/module/aim/action/AddAmpActivity.java
@@ -238,7 +238,7 @@ public class AddAmpActivity extends Action {
 							if (prevSelSectors != null) {
 								if (prevSelSectors.isEmpty()) {
 									selectedSector
-											.setSectorPercentage(new Integer(
+											.setSectorPercentage(new Float(
 													100));
 								} else {
 									selectedSector.setSectorPercentage(null);
@@ -251,7 +251,7 @@ public class AddAmpActivity extends Action {
 								}
 								prevSelSectors.add(selectedSector);
 							} else {
-								selectedSector.setSectorPercentage(new Integer(
+								selectedSector.setSectorPercentage(new Float(
 										100));
 								prevSelSectors = new ArrayList<ActivitySector>();
 								prevSelSectors.add(selectedSector);
@@ -323,7 +323,7 @@ public class AddAmpActivity extends Action {
 					if (prevSelSectors != null) {
 						if (prevSelSectors.isEmpty()) {
 							selectedSector
-									.setSectorPercentage(new Integer(100));
+									.setSectorPercentage(new Float(100));
 						} else {
 							selectedSector.setSectorPercentage(null);
 							Iterator iter = prevSelSectors.iterator();
@@ -335,7 +335,7 @@ public class AddAmpActivity extends Action {
 						}
 						prevSelSectors.add(selectedSector);
 					} else {
-						selectedSector.setSectorPercentage(new Integer(100));
+						selectedSector.setSectorPercentage(new Float(100));
 						prevSelSectors = new ArrayList<ActivitySector>();
 						prevSelSectors.add(selectedSector);
 					}
@@ -448,10 +448,10 @@ public class AddAmpActivity extends Action {
       if (addComponente) {
 			if (prevSelComponentes != null) {
 				if (prevSelComponentes.isEmpty())
-					selectedComponente.setSectorPercentage(new Integer(100));
+					selectedComponente.setSectorPercentage(new Float(100));
 				prevSelComponentes.add(selectedComponente);
 			} else {
-				selectedComponente.setSectorPercentage(new Integer(100));
+				selectedComponente.setSectorPercentage(new Float(100));
 				prevSelComponentes = new ArrayList<ActivitySector>();
 				prevSelComponentes.add(selectedComponente);
 			}
